Safari and Firefox Support
- Fixes issue where Safari and Firefox couldn't load rhymes due to a CORS
issue with the rhymes API.
- Fixes issue where Firefox would fail to display the "Open File" dialog due
to the setTimeout being done on listener calls losing the fact that it was
caused by a user-initiated action.
- Fixes issue where Safari didn't have the Blob.text function, causing file
opens to fail.
- Fixes issue where the download button would disappear when the Menu
component reloaded.

As apart of this, the code that forced web listeners to be invoked async
is removed. This had the consequence of forcing fixes for a few bugs where
the App component would say that it was ready for events before all event
listeners were actually ready, which would leave those events lost until
app did another ready-for-events.

The fix for that is to instead wrap the whole app in a component specifically
for calling ready-for-events. Since React useEffects hook are child first,
then parent, by having the last parent send ready-for-events, we can be
certain that all other listeners will be registered before ready-for-events
is sent.