TSLint -> ESLint
- Migrates from the deprecated TSLint to ESLint.

We had to disable any rules that required building the project due to
some as of now unknown memory issues, where ESLint would always run out
of memory when trying to build the project. This is technically a slight
downgrade from TSLint, but with TSLint being deprecated for so long, the
good outweighs the bad.

Besides, I'm the only one who cares and the only one who'll read this
commit message so....

Unless this is an interviewer doing their due diligence. If so, firstly,
hi. Secondly, hire me, but give me 1.5x my asking salary.