Migrates from Bazel to some homegrown Gulp tasks.
- Wardell, Wardell, Wardell. Why did you do this? Are you reading this
for some hope of understanding what you actually did here? Too bad,
'cause for some wild reason, you decided to do half of this 1 month and
a half ago and finish the rest of it at 1am after a very long week of
work. All you know is that the tests pass and Electron and Web seem to
be building. If you were hoping for more info, sorry. Just know that
Past Wardell is even angrier at Past-er Wardell for putting us in this
 situation. Good luck. You obviously need it way more than I do right
 now.