ESLint - Prettier - Back with a Vengenance
- Reenables Prettier running as apart of the lint rule and reformats
some files that were in violation.

I believe this was missed when we did the migration from TSLint to
ESLint, and we didn't notice for so long due to IDEA's built-in Prettier
integration mostly masking this. However, quite a lot of files ended up
being in violation regardless, so it's good to reenable it.