Renderer - Adds logic to immediately call listeners set by the Renderer
- Historically, when the renderer adds a listener to the platform, the
renderer either also sends some sort of "request" event or waits for the
"ready-for-events" to fire. Adding in a new "request" event is extra
work and waiting for the "ready-for-events" can be error-prone. To solve
both of those problems, now the platform can register a listener that
will be invoked whenever the renderer adds a new listener to a channel,
which will allow the platform to immediately update the renderer with
whatever data it wants. This commit uses this new feature to modify the
PreferenceManager to no longer use "ready-for-events". Instead, whenever
the renderer registers a "prefs-updated" or a "dark-mode-toggled"
listener, the PreferenceManager immediately sends its current knowledge
of what the preferences or theme are, respectively.