Fixes failing Electron UI tests
- Adds a new Typescript transformer to automatically add displayName to
functional components.
- Regenerates the package-lock to pull in a new version of resq, which
is used by WebdriverIO to find React components, which has a fix in it
to prefer display name, which is what was failing this test as Webpack
started changing the name of the Menu component to menu_Menu_Menu in
builds.
- Enables Dev Tools on UI test builds.
- Adds React Dev Tools as a dev dependency and includes instructions on
how to use it.
- Starts using TSLib to fix an issue with the
undefinedundefinedundefined bug appearing again.