feat: add in a new concept of "File history"
File History is the idea of storing the history of the lyrics as they
change so that eventually, we can allow users to arbitrarily load a
previous version of their lyrics. This also switches the
UnsavedDataManager to use File History to restore unsaved data as
opposed to its previous method of saving lyrics directly to temporary
files.