refactor: remove EditorTextStore and SelectedWordStore
These "stores" were heavily complicating testing, as they were hiding
the logic for knowing the selected text and the editor text. That logic
has no been moved into the App component, and the Editor and Rhyme
components have been updated to be controlled components. Well, as much
as possible; Editor still manages the CodeMirror state itself but it
will always display the text that App passes into it. Which also makes
testing text changes much easier.

Outside of that, the word replacement logic now lives entirely in the
App component as well, instead of half-way between the SelectedWordStore
and the CodeMirror component.

The CodeMirror component has also been re-written to be a React hook to
make it easier to use in a more controlled fashion, as now it can return
functions that allow the Editor component to more directly modify its
state without needing to actually know what that state is.