refactor(electron): don't remove listeners that are already removed
Because of how the Renderer and Platform delegates work for Electron,
when a listener has been removed, if we try to remove it again, we end
up passing in "null" as a value to Electron's ipcMain and ipcRenderer,
which will throw an error. Instead, manually check if we've already
removed a listener and don't try and remove it again.
Also, fix showRendererDialog to not try and remove listeners multiple
times.