fix(electron): support updates when latest release isn't for electron
Before this, if the latest available release on GitHub was for a version
that didn't have an Electron release associated with it, Electron
wouldn't allow updating to any version, as it would just always pull
the latest release instead of looking for the latest Electron release.
Since electron-updater can't be concerned with the intricacies of the
Lyricistant release process, we instead handle all of that logic
ourselves by querying the GitHub API directly, finding the latest
Electron release, and feeding the URL to download the app from into
electron-updater.

Also rearranges the electron tests so that the quick unit tests run
before we start bundling the app.