fix: saving a file for the first time will allow re-saves
There was a bug where we weren't setting the file handler used to save a
file but we WERE setting the path for that file. Due to this, when a
file was saved for the first time and we attempted to save it again, we
would get a NPE due to us expecting that the file handler wouldn't be
null. This fixes that, and also changes how the File Manager stores the
current file path & file handler by more tightly coupling them. This
makes it more difficult for them to be out of sync in the future.

This ALSO fixes a bug where the first save of a file wouldn't properly
persist the FileHandle in Chrome, meaning the "Save As" dialog would
display every single time you attempted to save until you opened the
file via the "Open" dialog.