fix(web): allow downloads to the latest available version
Before this fix, the latest release URL was constant, which was fine
when all releases would always have all platforms. However, if the
latest release was only for non-desktop platforms, then attempting to
download would take you to an incorrect URL. We fix this the same way
we fixed the Electron issue: query the GitHub API to get the latest
release we can download and instead link to that.