fix(web): stop showing "couldn't open file" on file open cancel
We were attempting to access the ArrayBuffer from fileOpen without
checking if it was null or not, which would throw and get caught by the
FileManager's error handling. That's not really an error though, so
instead just check if the result is null or not before trying to access
the ArrayBuffer.