test: reduce flakes by waiting for elements to appear
queryBy is nice because it's quick, but it doesn't wait at all. If we
use the findBy methods instead, the test will wait for the expectation
to pass based on the timeout, and then fail if it still hasn't appeared.
This should help to reduce the UI test flakes since CI will also wait
longer for findBys to become true.