feat: add in dictionary look-ups when typing lyrics
TabbedPane is dead. Long live DetailPane! Changed in order to keep both
the Rhymes and Dictionary components rendering even when they aren't
visible so that state isn't lost. Anyway, onto the feature:

So we've got a new Dictionary component that will show the dictionary
definitions for the current "query", where query is defined as either
the word the user has highlighted or the word nearest to the cursor.
This will always be the same word as the word we've fetching rhymes for.

The rest of these changes are moving some UI code that previously Rhymes
was using exclusively into a shared place for that Rhymes new sibling,
Dictionary, can access it too. Awh, don't they look alike? So cute!