refactor: make detail panes swipable on mobile
Just a better experience. This also adds in functionality not to load
rhymes or a definition when that panel is not being actively displayed
because we appreciate our APIs. Also reverts the change to only show
loading indicators when we're actively doing working and now shows them
when we're debouncing too. Just looks a bit nicer.