perf: improve fps when actively typing
We were constantly recreating a ton of event handler purely to access
the latest state, which caused a lot of work to be done on keypress.
More importantly, we were also constantly converting the Editor text's
from CodeMirror's format to string on every keypress, which while useful
in reducing the CodeMirror dependency, was hurting performance majorly
as that conversion gets more and more expensive the more text is
written.

The solution here is two-fold:

1. Use @mui's useEventCallback whereever possible. It's not as good as
the future useEvent from React proper, but it works for our purposes. It
lets us declare an event handler that can see the latest state but not
need to be recreated whenever that dependency changes.
2. Stop converting the Editor's text to string before putting it in
App's state. Instead, do the string conversion as late as possible;
pretty much right when we're about to send it to the platform or
perform a text replacement. This forces the App and Editor components to
know and care about what Text (CodeMirror's name for its textual
representation) is but the performance improvement is noticeable, so
it's a worthy trade-off here.

For text replacement specifically, we still bother to deal with the
entire document as a string instead of as a CodeMirror Text as Text has
its own ideas about what constitutes an "empty" string and its
line-based approach makes everything more complex than it needs to be.
Text.replace isn't quite the one-line replacement it might seem when
dealing with replacements in the middle of a Text document. It also
keeps this change as small as possible even though there's likely
improvements to be made there as well.