refactor: have detail pane manage the loading indicators for children
In a MUCH cooler fashion, the Detail Pane will now animate the selected
tab indicator whenever something in the detail pane is loading
something. This also gets rid of the padding around the Detail Pane (to
let the tabs be truly full width) and reverts the revert of making
loading indicators show even when debouncing.

Okay, because I know I'm going to be confused as to why I keep going
back and forth on this loading indicator business, I feel that when the
loading indicator is the "disabled" color, it can show all the time
because it isn't distracting. However, when it's the primary color, it
should only display when we're actually loading something because the
primary color draws the eye's attention. Since this is back to being the
primary color, we should only show it when not debouncing.