refactor: move cursor to always be after the replaced text
This was previously a bit random; it would work on single words but end
up highlighting all text if there was already a highlighted selection.
This is a bit better since it'll just place the cursor after the last
word but could be improved in the highlighted selection case, but hey,
what's good for the goose is good for the gander. Don't think I'm using
that phrase right, but I'm the only one who reads these anyway.