perf: improve performance when typing by reducing work on every keypress
The RhymesList is a very expensive component to render since it have hundreds of children.
Previously, it would render constantly as due to the lack of React.memo and non-memoized callbacks,
React would think it always needed to be updated. By memoizing as much as possible, including the
component itself, we reduce the amount of re-renders React does when the text changes. We also
change WordChip to not use sx since having so many of those caused Emotion to slow renders down a
bit even though WordChip was using a pretty simple style. We also copy these changes over to
Dictionary even though it wasn't performing that poorly since they share much of their structure.