ci: bump GitHub Action versions
Bumps all GitHub Actions to close-to-latest versions. Also removes an
outdated Automatic Releases action. This required changing how we handle
changelog generation for nightly builds, since now we have to do that
ourselves. That's not entirely true but better to have full control over
it, so we can tweak it in the future instead of relying on GitHub's.