fix(web): don't crash when cancelling saving logs
Deja vu here, but this should finally squash this class of issues. What
Files implementations should do when the user cancels hadn't been firmly
locked down, but this makes sure they all know to throw an exception
when things get bad and to return null when users cancel. This also
changes runIgnoringAbortErrors to fulfill that contract by returning
a "cancelled" flag with its response so that consumers can easily tell
that it was cancelled by the user and not that the function they called
returned null.