feat: add ability to hide the detail pane (rhymes and dictionary)
The detail pane, which holds the Rhymes and Dictionary functionality,
can now be toggled away, leaving more screen estate. There is also a new
setting that can be changed to disable this feature. On mobile, the new
toggle functionality will default to on, while other platforms will
default to off.

Co-authored-by: DJ <djsm00thee3@gmail.com>