refactor: keep React root mounted for the inspiration button
There's seemingly a race condition sometimes where we attempt to unmount
the React root we use for the Inspiration button after CodeMirror has
destroyed the DOM components. Instead of completely destroying the
root as well, we instead now keep it around and simply force it to
render "null" when CodeMirror tells us the widget has been removed and
instead remove the root once the plugin has been destroyed.