feat: support comments, todo, and section markers in lyrics
This uses the "//" style of comments, where starting a line with //
designates it a comment and any text after // designates that section
of the line as a comment. Having a comment that looks like "// todo"
(optional spacing and capitalization) will make a TODO comment instead,
which behaves the same but is more visually striking. Section markers
are in the form of [section] where section can be any text but the
marker has to be alone on a line for it to be properly recognized.