feat: add the new Lyricistant logo and change the theme colors
Adds in the adorable new Lyricistant logo for all platforms and
additionally adds support for themed icons on Android. Which was
actually the impetus for the new icon in the first place; I couldn't
quite figure out how to get the feather to look nice when monochrome.
Since the new logo uses an orange as its default color, this also
changes the primary color to be complimentary of the new logo.