fix: allow highlighting text starting from the end of the selection
A previous bug fix to put the cursor at the end of a text replacement
broke the ability to highlight text when starting from the end of a
selection and moving to the left. Highlighting still worked when going
from start to end. This fix just changes the logic to move CodeMirror's
internal "selection" to match the app's known cursor position when the
current selection is empty. This will allow users to highlight as normal
while still keeping the previous bug fix working since text replacements
always end up with CodeMirror having an internal selection that is
empty.