ci: add pause after tag deletion before creating release
Sometimes the actual deletion of the tag happens a bit after a new
release is created, leading to a race condition that leaves the Nightly
release in a draft state since it's created with a tag that stopped
existing. This should stop that from happening by waiting 30 seconds
after deleting a tag before creating a new release.