ci: use Java 17 for mobile builds and update Ruby for macOS 13
Android now requires Java 17 (instead of Java 11) to build so set up CI
to use that for mobile builds. iOS won't care about it so no need to
gate it for only Android.
Fastlane still requires Ruby, but we needed to update the setup Ruby
action to work on macOS 13, which iOS now requires.