refactor(electron): stop sorting electron logs
Sorting naively like this resulted in the logs being wildly out of order
due to the fact that many logs include lines that are only parts of the
log as a whole. This is common for any logs that include args that are
formatted to JSON. This does mean that now log messages will be out of
order relative to when they were sent (i.e., main logs will be in one
section and renderer logs in another) but that can still be solved
when viewing logs manually. The effect of calling "sort" can't.