refactor: show related text for words as well as meanings
Previously, we only shows the related text (synonyms & antonyms) for
definitions that included those in the specific meanings. Turns out
that we also get that data for the word itself. This shows the related
texts that we get for a full word. Since those can have duplicates,
we also remove any before displaying. This also slightly tweaks the UI
to show numbers between the different meanings to make it more clear
what's being displayed.