feat(ios): support setting a file name before saving
This changes how file saving works on iOS in a way that allows us to
keep a handle to the file that's been saved. This finally brings iOS
up to the other platforms so that now, saving a file once means you
can KEEP saving to that file, instead of having to continuously export
it.